refactors telepathy audio crate
